netwerk/test/mochitests/test_idn_redirect.html

<!DOCTYPE HTML>
<html>
<!--
  Bug 1142083 - IDN Unicode domain redirect is broken
  This test loads redirectme.html which is redirected simple_test.html, on a different IDN domain.
  A message is posted to that page, with responds with another.
  Upon receiving that message, we consider that the IDN redirect has functioned properly, since the intended page was loaded.
-->
<head>
  <title>Test for URI Manipulation</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<pre id="test">
<script type="text/javascript">

SimpleTest.waitForExplicitFinish();

var iframe = document.createElement("iframe");
iframe.src = "about:blank";
iframe.addEventListener("load", finishTest);
document.body.appendChild(iframe);
iframe.src = "http://mochi.test:8888/tests/netwerk/test/mochitests/redirect_idn.html";

function finishTest(e) {
  ok(true);
  SimpleTest.finish();
}

</script>

</body>
</html>
